examples/ipsec-secgw: skip Rx scatter for HW reassembly
When HW reassembly is enabled, currently both Rx scatter and
Tx multi-segs ethdev offload flags are enabled. Tx multi-seg is
needed to Tx the reassembled pkt but Rx scatter is not needed as
reassembly should be self sufficient offload flag. Hence remove it
and only enable Tx multi-segs when HW reassembly is enabled.

Signed-off-by: Nithin Dabilpuram <ndabilpuram@marvell.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>