net/sfc: add support for IPv4 NAT offload to MAE backend
For this offload to work, the innermost pattern items must
provide the full set of exact match criteria, which are as
follows: EtherType, IP DST, IP SRC, TP protocol ID, TP DST
and TP SRC, where the protocol types can be autodetected.

The offload requires that the IPv4 and the TP actions be
requested simultaneously in the same flow by the caller:
SET_IPV4_DST + SET_TP_DST or SET_IPV4_SRC + SET_TP_SRC.

The offload operates on the outermost frame, which,
if action VXLAN_DECAP was requested, maps to the
inner frame of the original packet. The caller
is responsible to request this offload only
when the target header is an IPv4-based one.

Signed-off-by: Ivan Malov <ivan.malov@arknetworks.am>
Reviewed-by: Andy Moreton <amoreton@xilinx.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>