buildtools: remove assumption about current work dir
According to meson documentation, we should not make any assumptions as
to what directory a script is called from[1]. The use of "relpath" when
processing the output of directory globbing is therefore unsafe, as it
was stripping off paths relative to the current directory - which just
happened to be the same as the root directory we were processing.

To improve safety, and support meson clones (like muon)  which
don't always run things from the currently processed path, we need to
explicitly specify that the start parameter for relpath() should be
"root" value.

[1] https://mesonbuild.com/Reference-manual_functions.html#run_command

Signed-off-by: Bruce Richardson <bruce.richardson@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>