net/nfp: fix getting firmware version
The VF PMD can not get a valid 'cpp' data filed in the 'struct
nfp_net_hw', the value of which will be NULL always. This cause any
logic which use of 'cpp' will make the app crash.

Fix this by delete the support of getting firmware version of VF.

Fixes: 128c8ad951bf ("net/nfp: support getting firmware version")
Cc: chaoyong.he@corigine.com

Signed-off-by: Chaoyong He <chaoyong.he@corigine.com>
Reviewed-by: Long Wu <long.wu@corigine.com>
Reviewed-by: Niklas Söderlund <niklas.soderlund@corigine.com>
Reviewed-by: Peng Zhang <peng.zhang@corigine.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>