net/nfp: changes to link status reporting
The NFP PMD driver retrieves the linkspeed by reading the
NFP_NET_CFG_STS_LINK register. This register is not set by all versions
of the firmware. Changes in this patch include:

Addition of listener function to deal with control messages of type
NFP_FLOWER_CMSG_TYPE_PORT_MOD being received by the PMD. This ensures
that the up/down status of the port is updated when a version of
firmware is being used that does not write the link status to the
NFP_NET_CFG_STS register.

Update the nfp_flower_repr_link_update() function to preferably read the
linkspeed from the port table, while keeping the old lookup method as a
fallback in case that is not supported.

Signed-off-by: James Hershaw <james.hershaw@corigine.com>
Reviewed-by: Long Wu <long.wu@corigine.com>
Reviewed-by: Chaoyong He <chaoyong.he@corigine.com>
Reviewed-by: Niklas Söderlund <niklas.soderlund@corigine.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>