crypto/mlx5: add enqueue and dequeue operations
The crypto operations are performed with crypto WQE. If the input
buffers(AAD, mbuf, digest) are not contiguous and there is no enough
headroom/tailroom for copying AAD/digest, as the requirement from FW,
an UMR WQE is needed to generate contiguous address space for crypto
WQE. The UMR WQE and crypto WQE are handled in two different QPs.

Crypto operation with non-contiguous buffers will have its own UMR
WQE, while the operation with contiguous buffers doesn't need the
UMR WQE. Once the all the operations WQE in the enqueue burst built
finishes, if any UMR WQEs are built, an additional SEND_EN WQE will
be as the final WQE of the burst in the UMR QP. The purpose of that
SEND_EN WQE is to trigger the crypto QP processing with the UMR ready
input memory address space buffers.

The QP for crypto operations contains only the crypto WQE and the QP
WQEs are built as fixed in QP setup. The QP processing is triggered
by doorbell ring or the SEND_EN WQE from UMR QP.

Signed-off-by: Suanming Mou <suanmingm@nvidia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>