net/mlx5: support bus socket with no hugepages
When a Mellanox NIC is attached to a bus on a numa
socket, it tries to allocate rte memory in that socket.
If hugepages are not configured/available on that rte socket
mlx5_common_pci_probe fails with ENOMEM.

In this patch, a memflag MLX5_MEM_FALLBACK_ANY_SOCKET is
introduced which when set on mlx5_malloc, will allocate
the memory using SOCKET_ID_ANY if it is not able to allocate
memory on the specified socket. This allocates memory on
any socket starting with the current thread's socket.

Signed-off-by: Hari Sasank <harisasank@outlook.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>