regex/cn9k: remove rule compiler
Nobody knows how to build the feature.
When the dependency "rxp_compiler" is found,
the header file is not available:

drivers/regex/cn9k/cn9k_regexdev_compiler.c:12:10: fatal error:
	rxp-compiler.h: No such file or directory

It seems that it depends on a proprietay library.
According to what was discussed in the Technical Board:
http://mails.dpdk.org/archives/dev/2019-June/135847.html
the dependency must be "freely available".

Bugzilla ID: 1232

Signed-off-by: Thomas Monjalon <thomas@monjalon.net>
Signed-off-by: 0-day Robot <robot@bytheb.org>