event/dlb2: skip configuration if no eventdev lib
While the build system will skip building most libs and drivers when a
dependency is missing for a component, for DLB2 driver, the
"static_rte_eventdev" object is referenced inside the meson.build file
itself, which will cause crashes if it doesn't exist i.e. if eventdev is
disabled. Prevent this issue by skipping processing the file if no
eventdev. [The build system will still report missing dependency, as the
dependency is set by default for all eventdev drivers]

Signed-off-by: Bruce Richardson <bruce.richardson@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>