net/ice: revert mark action in switch filter
For current implementation, switch filter missed the consumption part of
mark id for app, and switch filter captures the rules first instead of
FDIR, which makes the FDIR rules lost.

Revert this commit to disable mark action in switch filter and recover
the mark action hit in FDIR by default.

Fixes: e24f43f2c824 ("net/ice: support flow mark action in switch")
Signed-off-by: Zhichao Zeng <zhichaox.zeng@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>