pci: add O_CLOEXEC when open uio device
In this scenario, the DPDK process invokes a script which
inherits an open file descriptor (FD) for a UIO device.
After the script execution is complete, the UIO device's
close operation is called. However, in a new kernel version
(865a11f987ab5f03:uio/uio_pci_generic: Disable bus-mastering on release),
this close operation causes the PCI bus master bit to be cleared,
rendering the device unusable and leading to unexpected behavior.

This modification was made to prevent the UIO device's FD
from being inherited by the child process.

Signed-off-by: Weifeng Su <suweifeng1@huawei.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>