docs: freebsd: Update to 20.11
This patch updates the installation instructions for DPDK on FreeBSD.
It specifies the explicit version of DPDK (20.11) to be installed.
This change is important as the 'dpdk' package is an alias and doesn't
always point to the latest version. By specifying the explicit version,
we make it clear which version is to be installed. The page previously
showed 'pkg install dpdk' without specifying the version.

Signed-off-by: David Young <dave@youngcopy.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>