set namespace prefix to threads
When looking at threads in a system, it can be confusing
to find some unknown threads without a clue it is started by DPDK.

Let's start all thread names with "dpdk-"
and the driver name if it comes from a driver.

One more constraint: the thread names are generally limited
to 16 characters, including NUL character.

Signed-off-by: Thomas Monjalon <thomas@monjalon.net>
Signed-off-by: 0-day Robot <robot@bytheb.org>