bus/vdev: fix memory leak
In hotplug usecase, devargs will be allocated in secondary process
in the function alloc_devargs. Since it will not be insert into the
devarg_list, it will have no chance to be freed.

This patch adds additional memory free for device structure member devargs
in the secondary process in rte_vdev_uninit.

Fixes: f5b2eff0847d ("bus/vdev: fix devargs after multi-process bus scan")
Cc: stable@dpdk.org

Signed-off-by: Wenjun Wu <wenjun1.wu@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>