eal: unify logging code
FreeBSD logging code was not using syslog and did not have
the same options as Linux. Use a common set of functions for that.

Pre-parsing for log level is common to all OS's.

Now the pre-scan can return an error.
A bad argument give to --log-level option was given the
code would keep going.

Use opterr to fix duplicate error message on bad option.

Signed-off-by: Stephen Hemminger <stephen@networkplumber.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>