eal: allow both allow and block options coexistence
Currently, all buses use the same eal allow and block options. Need to
allow both allow and block options for different buses to coexist.
It wouldn't be a problem for pci bus if both allow and block options
were present. When the first option occurs, the scan mode for pci bus is
set.

For example:
  --allow 0000:05:00.0 --block wq0.0
    only pci device 0000:05:00.0 will be scanned
    all devices except wq0.0 on dsa bus will be scanned
  --allow 0000:05:00.0 --block 0000:05:00.1
    block option will be ignored
  --block 0000:05:00.1 --allow 0000:05:00.0
    allow option will be ignored

Signed-off-by: Marvin Liu <yong.liu@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>