lib/ethdev: sanity check on callback before resource allocation
sanity check is performed after a rte_eth_dev object is allocated.
The object is not freed if the check fails, though in the current
reality this never happens, but it's better programming paradigm
to move the quick check up front to the start of
rte_eth_dev_pci_generic_probe.

Signed-off-by: Kaiyu Zhang <squirrel.prog@gmail.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>